+++ ui/pages/create-account/import-account/bottom-button.stories.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import BottomButtons from './bottom-buttons';
+
+export default {
+  title: 'Pages/CreateAccount/ImportAccount/BottomButtons',
+  argTypes: {
+    importAccountFunc: {
+      control: {
+        type: 'function',
+      },
+    },
+    isPrimaryDisabled: {
+      control: {
+        type: 'boolean',
+      },
+    },
+  },
+};
+
+export const DefaultStory = (args) => <BottomButtons {...args} />;
+
+DefaultStory.storyName = 'Default';
