+++ ui/pages/create-account/import-account/json.js
@@ -2,17 +2,23 @@ import PropTypes from 'prop-types';
 import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
 import FileInput from 'react-simple-file-input';
-import { ButtonLink, Text } from '../../../components/component-library';
-import FormField from '../../../components/ui/form-field';
+import {
+  ButtonLink,
+  TEXT_FIELD_SIZES,
+  TEXT_FIELD_TYPES,
+  Text,
+  FormTextField,
+} from '../../../components/component-library';
+
 import {
   Size,
   TextVariant,
   TEXT_ALIGN,
 } from '../../../helpers/constants/design-system';
+import ZENDESK_URLS from '../../../helpers/constants/zendesk-url';
 import { useI18nContext } from '../../../hooks/useI18nContext';
 import { displayWarning } from '../../../store/actions';
 import BottomButtons from './bottom-buttons';
-import { moreInfoLink } from '.';
 
 JsonImportSubview.propTypes = {
   importAccountFunc: PropTypes.func.isRequired,
@@ -45,7 +51,12 @@ export default function JsonImportSubview({ importAccountFunc }) {
     <>
       <Text variant={TextVariant.bodyMd} textAlign={TEXT_ALIGN.CENTER}>
         {t('usedByClients')}
-        <ButtonLink size={Size.inherit} onClick={moreInfoLink}>
+        <ButtonLink
+          size={Size.inherit}
+          href={ZENDESK_URLS}
+          target="_blank"
+          rel="noopener noreferrer"
+        >
           {t('fileImportFail')}
         </ButtonLink>
       </Text>
@@ -55,17 +66,20 @@ export default function JsonImportSubview({ importAccountFunc }) {
         onLoad={(event) => setFileContents(event.target.result)}
         style={{
           padding: '20px 0px 12px 15%',
-          fontSize: '15px',
+          fontSize: '16px',
           display: 'flex',
           justifyContent: 'center',
           width: '100%',
         }}
       />
 
-      <FormField
+      <FormTextField
+        size={TEXT_FIELD_SIZES.LARGE}
         autoFocus
-        password
-        error={warning}
+        id="json-password-box"
+        type={TEXT_FIELD_TYPES.PASSWORD}
+        helpText={warning}
+        helpTextProps={{ error: true }} // TODO: change to severity={SEVERITIES.ERROR} after rebase
         placeholder={t('enterPassword')}
         id="json-password-box"
         value={password}
@@ -73,8 +87,8 @@ export default function JsonImportSubview({ importAccountFunc }) {
         inputProps={{
           onKeyPress: handleKeyPress,
         }}
+        marginBottom={8}
       />
-
       <BottomButtons
         importAccountFunc={_importAccountFunc}
         isPrimaryDisabled={isPrimaryDisabled}
