+++ ui/pages/create-account/import-account/private-key.stories.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import PrivateKeyImportView from './private-key';
+
+export default {
+  title: 'Pages/CreateAccount/ImportAccount/PrivateKeyImportView',
+  argTypes: {
+    importAccountFunc: {
+      control: {
+        type: 'function',
+      },
+    },
+  },
+};
+
+export const DefaultStory = (args) => <PrivateKeyImportView {...args} />;
+
+DefaultStory.storyName = 'Default';
